factorise common behaviors between QF APIs
